upbit/pyproject.toml
[project]
name = "upbit"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "mcp>=1.8.0",
    "numpy>=2.2.5",
    "pandas>=2.2.3",
    "python-dotenv>=1.1.0",
]
